Refactor modules
