+++ src/hooks/dx11.rs
@@ -21,8 +21,8 @@ use windows::Win32::Graphics::Dxgi::{
 };
 
 use super::DummyHwnd;
-use crate::hooks::render::RenderState;
 use crate::mh::MhHook;
+use crate::renderer::RenderState;
 use crate::{Hooks, ImguiRenderLoop};
 
 type DXGISwapChainPresentType =
