+++ src/hooks/dx9.rs
@@ -13,8 +13,8 @@ use windows::Win32::Graphics::Direct3D9::{
 use windows::Win32::Graphics::Gdi::RGNDATA;
 
 use super::DummyHwnd;
-use crate::hooks::render::RenderState;
 use crate::mh::MhHook;
+use crate::renderer::RenderState;
 use crate::util::try_out_ptr;
 use crate::{Hooks, ImguiRenderLoop};
