+++ src/hooks/mod.rs
@@ -34,9 +34,7 @@ use windows::Win32::UI::WindowsAndMessaging::{
 pub mod dx11;
 pub mod dx12;
 pub mod dx9;
-mod input;
 pub mod opengl3;
-mod render;
 
 pub fn find_process_hwnd() -> Option<HWND> {
     static mut FOUND_HWND: OnceLock<HWND> = OnceLock::new();
